fixxx

+++ dsipts/models/Diffusion.py
@@ -582,7 +582,7 @@ class SubNet(nn.Module):
 
         import pdb
         pdb.set_trace()
-        emb_y_noised = self.y_noised_linear(y_noised)
+        emb_y_noised = self.y_noised_linear(y_noised.float())
         emb_y_past = self.y_past_linear(y_past)
         
         # LIN FOR PAST
